refactor: Extracting formatting code
